Update userAuth.py
The invite to the discord was not being presented at the right time. I moved it so that after it fails to get any roles (meaning they are not part of the server yet) they will be directed to join the server.

I also changed the log level of people authenticating so you can see when someone fails any possibly assist them.